requirements.txt
absl-py==1.4.0
ansi2html==1.8.0
array-record==0.2.0
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
auditwheel==5.3.0
backcall==0.2.0
cachelib==0.9.0
cachetools==5.3.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.10.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-extensions==1.0.1
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.10.2
dash-table==5.0.0
dash-tools==1.11.1
dask==2023.3.1
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
Django==4.2.1
dm-tree==0.1.8
docopt==0.6.2
EditorConfig==0.12.3
einsum==0.3.0
etils==1.2.0
executing==1.2.0
filelock==3.12.0
Flask==2.2.3
Flask-Caching==2.0.2
flatbuffers==23.5.8
fonttools==4.39.3
fsspec==2023.5.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
gunicorn==20.1.0
h5py==3.8.0
identify==2.5.24
idna==3.4
imageio==2.28.1
immutabledict==2.2.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsbeautifier==1.14.7
jupyter-dash==0.4.2
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
locket==1.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.1
matplotlib-inline==0.1.6
more-itertools==9.1.0
nest-asyncio==1.5.6
networkx==3.1
nodeenv==1.8.0
numpy==1.21.5
oauthlib==3.2.2
OpenEXR==1.3.9
opt-einsum==3.3.0
orjson==3.9.0
packaging==21.0
pandas==1.5.3
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==3.5.1
plotly==5.13.1
plotly-resampler==0.8.3.2
pluggy==1.0.0
pre-commit==2.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyelftools==0.29
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pytest==6.2.5
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.11.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.0
tensorflow-probability==0.19.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tk==0.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
trace-updater==0.0.9.1
traitlets==5.9.0
trimesh==3.21.6
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==2.0.2
virtualenv==20.23.0
visdcc==0.0.50
waymo-open-dataset-tf-2-11-0==1.5.1
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
yarg==0.1.9
yolk3k==0.9
zipp==3.15.0
